fixes to admission page
now with a bit better of a sidebar with scrollspy to easily navigate
between sections
I want to stay away from having outer links on the side nav because
they're pretty nondescript as to what you're actually going to be
looking at.
todo: seperate out a bunch of stuff, make it look nicer, etc.